Fix signature of VCustomSource#SeekCallback (fixes #171)
Previously the offset was mistyped as a `*void`/`MemorySegment`. It was now
changed to be a `gint64`/`long`. Additionally, the `int whence`
parameter was replaced with a new `SeekWhence` enum that provides more
type safety and discoverability for implementers.

To test, a sample was added to provide a `VCustomSource` based on a
`FileHandle` that implements seeking, testing it against a TIF image.